move redux.js’ init logic into index.js’ init function  🐿 v2.9.0
